requirements.txt
cisco-ai-mcp-scanner>=3.2.0
requests>=2.31.0
rich>=13.7.0
python-dotenv>=1.0.0
httpx>=0.28.0
fastapi>=0.110.0
uvicorn>=0.29.0

